mcp-server/function_app.py
import azure.functions as func
import logging
import json
import os
import datetime
from typing import Dict, Any, List, Optional, Tuple
import time

# Import Microsoft Identity Web auth module
from azure_auth import authenticate_request, get_graph_token

# Create a function app
app = func.FunctionApp()

# Azure APIs will have their versions defined directly in each function
# This allows for more flexibility when different endpoints need different API versions

# Utility functions for response formatting
def create_error_response(status_code: int, message: str) -> func.HttpResponse:
    """Create a standardized error response."""
    return func.HttpResponse(
        body=json.dumps({"error": message}),
        mimetype="application/json",
        status_code=status_code
    )

def create_success_response(data: Any) -> func.HttpResponse:
    """Create a standardized success response."""
    return func.HttpResponse(
        body=json.dumps({"data": data}),
        mimetype="application/json",
        status_code=200
    )
    
    payload = verify_token(token)
    return payload

@app.route(route="authenticate", auth_level=func.AuthLevel.ANONYMOUS, methods=["POST"])
def authenticate(req: func.HttpRequest) -> func.HttpResponse:
    """Validate an Azure AD token and return user information."""
    try:
        # Authenticate the request
        is_authenticated, user_info, error_message = authenticate_request(req)
        
        if not is_authenticated:
            logging.error(f"Authentication failed: {error_message}")
            return create_error_response(401, f"Authentication failed: {error_message}")
        
        # Return user information to the client
        return create_success_response({
            "message": "Token validated successfully",
            "user_info": user_info
        })
        
    except Exception as e:
        logging.error(f"Authentication error: {str(e)}")
        return create_error_response(500, f"Authentication error: {str(e)}")

@app.route(route="incidents/list", auth_level=func.AuthLevel.ANONYMOUS, methods=["POST"])
def list_incidents(req: func.HttpRequest) -> func.HttpResponse:
    """List recent incidents from Azure Sentinel."""
    # API version specified directly in the function
    api_version = "2022-01-01-preview"  # Sentinel API version
    
    # Authenticate request using Azure AD
    is_authenticated, user_info, error_message = authenticate_request(req)
    
    if not is_authenticated:
        return create_error_response(401, error_message or "Unauthorized")
    
    try:
        req_body = req.get_json()
        limit = req_body.get("limit", 10)
        filter_params = req_body.get("filter", {})
        
        # In a production implementation, the Azure Sentinel API would be called here
        # This would use the Azure SDK to make the call with proper credentials
        
        # For now, we'll return an informative message that the API integration is pending
        return create_error_response(501, 
            "Azure Sentinel API integration pending. This endpoint would list "
            f"up to {limit} incidents with the filter: {filter_params}, "
            f"using API version: {api_version}")
    except Exception as e:
        logging.error(f"Error listing incidents: {str(e)}")
        return create_error_response(500, f"Error listing incidents: {str(e)}")

@app.route(route="incidents/get", auth_level=func.AuthLevel.ANONYMOUS, methods=["POST"])
def get_incident(req: func.HttpRequest) -> func.HttpResponse:
    """Get a specific incident by ID."""
    # API version specified directly in the function
    api_version = "2022-01-01-preview"  # Sentinel API version
    
    # Authenticate request using Azure AD
    is_authenticated, user_info, error_message = authenticate_request(req)
    
    if not is_authenticated:
        return create_error_response(401, error_message or "Unauthorized")
    
    try:
        req_body = req.get_json()
        incident_id = req_body.get("id")
        
        if not incident_id:
            return create_error_response(400, "Incident ID is required")
        
        # In a production implementation, the Azure Sentinel API would be called here
        # This would use the Azure SDK to make the call with proper credentials
        
        # For now, we'll return an informative message that the API integration is pending
        return create_error_response(501, 
            "Azure Sentinel API integration pending. This endpoint would retrieve the "
            f"incident with ID: {incident_id}, using API version: {api_version}")
    except Exception as e:
        logging.error(f"Error getting incident: {str(e)}")
        return create_error_response(500, f"Error getting incident: {str(e)}")

@app.route(route="metrics/dashboard", auth_level=func.AuthLevel.ANONYMOUS, methods=["POST"])
def get_metrics_dashboard(req: func.HttpRequest) -> func.HttpResponse:
    """Get security metrics dashboard data."""
    # API version specified directly in the function
    api_version = "2021-10-01"  # Defender API version for metrics
    
    # Authenticate request using Azure AD
    is_authenticated, user_info, error_message = authenticate_request(req)
    
    if not is_authenticated:
        return create_error_response(401, error_message or "Unauthorized")
    
    try:
        # In a production implementation, the Azure Defender API would be called here
        # This would use the Azure SDK to make the call with proper credentials
        
        # For now, we'll return an informative message that the API integration is pending
        return create_error_response(501, 
            "Azure Defender API integration pending. This endpoint would retrieve "
            f"security metrics dashboard data using API version: {api_version}")
    except Exception as e:
        logging.error(f"Error getting metrics dashboard: {str(e)}")
        return create_error_response(500, f"Error getting metrics dashboard: {str(e)}")

